update eslint and tsconfig

+++ src/services/reports/remove.ts
@@ -2,7 +2,13 @@ import getDevice from "@tago-io/sdk/out/modules/Utils/getDevice";
 import { RouterConstructorData } from "../../types";
 
 export default async ({ config_dev, context, scope, account, environment }: RouterConstructorData) => {
+  if (!account || !environment || !scope || !config_dev || !context) {
+    throw new Error("Missing parameters");
+  }
   const action_group = scope[0].group;
+  if(!action_group) {
+    throw new Error("Missing action group");
+  }
 
   const [action_registered] = await account.actions.list({
     page: 1,
@@ -13,7 +19,15 @@ export default async ({ config_dev, context, scope, account, environment }: Rout
     amount: 1,
   });
 
+  if(!action_registered.tags) {
+    throw new Error("Action not found in Tago");
+  }
+
   const org_id = action_registered.tags.find((x) => x.key === "organization_id")?.value;
+
+  if(!org_id) {
+    throw new Error("Organization not found in Tago");
+  }
 
   const org_dev = await getDevice(account, org_id);
   await org_dev.deleteData({ groups: action_group, qty: 9999 });
